# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hydra",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/transactions/": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Create Transaction",
        "description": "Record an expense, some income, or a transfer between accounts.\n\nA transfer is an ordinary transaction with kind \"transfer\" and a\ndestination account, so it needs no separate endpoint.\n\nArgs:\n    transaction_service: The transaction service dependency.\n    household: The current household context.\n    transaction_in: The transaction to record.\n\nReturns:\n    The recorded transaction.\n\nRaises:\n    HTTPException: If an account or category does not exist in the household\n        (404), or the transaction does not match its kind (400).",
        "operationId": "transactions-create_transaction",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "List Transactions",
        "description": "List the transactions of the household.\n\nFiltering by a parent category includes the spending filed under its\nsubcategories, unless include_subcategories is set to false.\n\nArgs:\n    transaction_service: The transaction service dependency.\n    household: The current household context.\n    filters: The date range, account, category, kind, amount range, search\n        text, ordering and paging to apply.\n\nReturns:\n    The matching transactions and the total number of matches.\n\nRaises:\n    HTTPException: If the category filter names a category outside the\n        household (404), or a filter is not recognised (422).",
        "operationId": "transactions-list_transactions",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "include_subcategories",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Subcategories"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TransactionKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "min_amount_minor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4611686018427387904,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Amount Minor"
            }
          },
          {
            "name": "max_amount_minor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4611686018427387904,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Amount Minor"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 255
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/TransactionSort",
              "default": "-date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionsPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Get Transaction",
        "description": "Get one transaction of the household.\n\nArgs:\n    transaction_service: The transaction service dependency.\n    household: The current household context.\n    transaction_id: The ID of the transaction.\n\nReturns:\n    The transaction.\n\nRaises:\n    HTTPException: If the transaction does not exist in the household (404).",
        "operationId": "transactions-get_transaction",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "transactions"
        ],
        "summary": "Update Transaction",
        "description": "Edit a transaction.\n\nArgs:\n    transaction_service: The transaction service dependency.\n    household: The current household context.\n    transaction_id: The ID of the transaction to edit.\n    transaction_in: The fields to change.\n\nReturns:\n    The updated transaction.\n\nRaises:\n    HTTPException: If the transaction, an account or a category does not\n        exist in the household (404), or the change does not match the kind\n        (400).",
        "operationId": "transactions-update_transaction",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transactions"
        ],
        "summary": "Delete Transaction",
        "description": "Delete a transaction.\n\nArgs:\n    transaction_service: The transaction service dependency.\n    household: The current household context.\n    transaction_id: The ID of the transaction to delete.\n\nReturns:\n    A confirmation message.\n\nRaises:\n    HTTPException: If the transaction does not exist in the household (404).",
        "operationId": "transactions-delete_transaction",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Message": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "Message"
      },
...
      "TransactionCreate": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TransactionKind"
          },
          "amount_minor": {
            "type": "integer",
            "maximum": 4.611686018427388e+18,
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "occurred_on": {
            "type": "string",
            "format": "date",
            "title": "Occurred On"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "counter_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Account Id"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "amount_minor",
          "occurred_on",
          "account_id"
        ],
        "title": "TransactionCreate"
      },
      "TransactionKind": {
        "type": "string",
        "enum": [
          "expense",
          "income",
          "transfer"
        ],
        "title": "TransactionKind"
      },
      "TransactionPublic": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TransactionKind"
          },
          "amount_minor": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor"
          },
          "occurred_on": {
            "type": "string",
            "format": "date",
            "title": "Occurred On"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "household_id": {
            "type": "string",
            "format": "uuid",
            "title": "Household Id"
          },
          "account_id": {
            "type": "string",
            "format": "uuid",
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "counter_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Account Id"
          },
          "recurring_rule_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurring Rule Id"
          },
          "is_generated": {
            "type": "boolean",
            "title": "Is Generated",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "amount_minor",
          "occurred_on",
          "id",
          "household_id",
          "account_id",
          "created_at"
        ],
        "title": "TransactionPublic"
      },
      "TransactionSort": {
        "type": "string",
        "enum": [
          "-date",
          "date",
          "-amount",
          "amount"
        ],
        "title": "TransactionSort"
      },
      "TransactionUpdate": {
        "properties": {
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "amount_minor": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.611686018427388e+18,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Minor"
          },
          "occurred_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred On"
          },
          "merchant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Merchant"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "counter_account_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Counter Account Id"
          }
        },
        "type": "object",
        "title": "TransactionUpdate"
      },
      "TransactionsPublic": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TransactionPublic"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "data",
          "count"
        ],
        "title": "TransactionsPublic"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login/access-token"
          }
        }
      }
    }
  }
}